move damage logging to utilities module
In the process, let's add some comments, improve some names,
and OH THE HORROR WHY DOES ECLIPSE MIX TABS AND SPACES BY DEFAULT?!
Better fix that too. And line endings?!
